Configure TracerFilter only in Spring MVC environments
Fixe sgh-154
