+++ tracer-spring-boot-starter/src/test/java/org/zalando/tracer/spring/FilterTest.java
@@ -8,25 +8,57 @@ import org.springframework.boot.autoconfigure.ImportAutoConfiguration;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.web.servlet.FilterRegistrationBean;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
-import org.springframework.test.context.web.WebAppConfiguration;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.notNullValue;
+import static org.hamcrest.Matchers.nullValue;
 
-@ExtendWith(SpringExtension.class)
-@SpringBootTest(classes = Application.class)
-@ImportAutoConfiguration(TracerAutoConfiguration.class)
-@WebAppConfiguration
 class FilterTest {
 
-    @Autowired
-    @Qualifier("tracerFilter")
-    private FilterRegistrationBean tracerFilter;
+    @ExtendWith(SpringExtension.class)
+    @SpringBootTest(classes = Application.class, properties = "spring.main.web-application-type=NONE")
+    @ImportAutoConfiguration(TracerWebMvcAutoConfiguration.class)
+    static class NonWebEnvironmentTest {
 
-    @Test
-    void shouldInitializeFilter() {
-        assertThat(tracerFilter, is(notNullValue()));
+        @Autowired(required = false)
+        @Qualifier("tracerFilter")
+        private FilterRegistrationBean tracerFilter;
+
+        @Test
+        void shouldNotInitializeFilter() {
+            assertThat(tracerFilter, is(nullValue()));
+        }
+    }
+
+    @ExtendWith(SpringExtension.class)
+    @SpringBootTest(classes = Application.class, properties = "spring.main.web-application-type=REACTIVE")
+    @ImportAutoConfiguration(TracerWebMvcAutoConfiguration.class)
+    static class WebFluxEnvironmentTest {
+
+        @Autowired(required = false)
+        @Qualifier("tracerFilter")
+        private FilterRegistrationBean tracerFilter;
+
+        @Test
+        void shouldNotInitializeFilter() {
+            assertThat(tracerFilter, is(nullValue()));
+        }
+    }
+
+    @ExtendWith(SpringExtension.class)
+    @SpringBootTest(classes = Application.class, properties = "spring.main.web-application-type=SERVLET")
+    @ImportAutoConfiguration(TracerWebMvcAutoConfiguration.class)
+    static class WebMvcEnvironmentTest {
+
+        @Autowired
+        @Qualifier("tracerFilter")
+        private FilterRegistrationBean tracerFilter;
+
+        @Test
+        void shouldInitializeFilter() {
+            assertThat(tracerFilter, is(notNullValue()));
+        }
     }
 
 }
